+++ Backend/routes/plantCatalog.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const mongoose = require("mongoose");
 require("dotenv").config();
-const build = require("../module/plantCatalogDBC");
+const build = require("../utils/plantCatalogDBC");
 
 
 router.get("/", (req, res) => {
